changed name to BrowserFavorite (#2)
