+++ Community.PowerToys.Run.Plugin.BrowserFavorite.Tests/Community.PowerToys.Run.Plugin.BrowserFavorite.Tests.csproj
@@ -7,6 +7,7 @@
     <Platforms>x64;ARM64</Platforms>
     <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
     <AppendRuntimeIdentifierToOutputPath>false</AppendRuntimeIdentifierToOutputPath>
+    <RootNamespace>Community.PowerToys.Run.Plugin.BrowserFavorite.Tests</RootNamespace>
   </PropertyGroup>
 
   <PropertyGroup>
@@ -22,7 +23,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Community.PowerToys.Run.Plugin.BraveFavorite\Community.PowerToys.Run.Plugin.BraveFavorite.csproj" />
+    <ProjectReference Include="..\Community.PowerToys.Run.Plugin.BrowserFavorite\Community.PowerToys.Run.Plugin.BrowserFavorite.csproj" />
   </ItemGroup>
 
 </Project>
