+++ Community.PowerToys.Run.Plugin.BrowserFavorite.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.33414.496
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Community.PowerToys.Run.Plugin.BraveFavorite", "Community.PowerToys.Run.Plugin.BraveFavorite\Community.PowerToys.Run.Plugin.BraveFavorite.csproj", "{18D111C3-66B7-49F6-A4AB-6485D903C1FA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Community.PowerToys.Run.Plugin.BrowserFavorite", "Community.PowerToys.Run.Plugin.BrowserFavorite\Community.PowerToys.Run.Plugin.BrowserFavorite.csproj", "{18D111C3-66B7-49F6-A4AB-6485D903C1FA}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Community.PowerToys.Run.Plugin.BraveFavorite.Tests", "Community.PowerToys.Run.Plugin.BraveFavorite.Tests\Community.PowerToys.Run.Plugin.BraveFavorite.Tests.csproj", "{D293A7C2-70BC-43F2-BAC3-1E43CD1DAE36}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Community.PowerToys.Run.Plugin.BrowserFavorite.Tests", "Community.PowerToys.Run.Plugin.BrowserFavorite.Tests\Community.PowerToys.Run.Plugin.BrowserFavorite.Tests.csproj", "{D293A7C2-70BC-43F2-BAC3-1E43CD1DAE36}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{90031F25-FF55-4164-B616-7D1450451ED5}"
 	ProjectSection(SolutionItems) = preProject
