+++ Community.PowerToys.Run.Plugin.BrowserFavorite/BrowserSource.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Davide Giacometti. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Community.PowerToys.Run.Plugin.BraveFavorite;
+namespace Community.PowerToys.Run.Plugin.BrowserFavorite;
 
 public enum BrowserSource
 {
