+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Images/Folder.dark.png
