+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Images/Folder.light.png
