+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Images/Url.dark.png
