+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Images/Url.light.png
