+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Libs/ARM64/PowerToys.Common.UI.dll
