+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Libs/ARM64/PowerToys.ManagedCommon.dll
