+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Libs/ARM64/PowerToys.Settings.UI.Lib.dll
