+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Libs/ARM64/Wox.Plugin.dll
