+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Libs/x64/PowerToys.Common.UI.dll
