+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Libs/x64/PowerToys.ManagedCommon.dll
