+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Libs/x64/PowerToys.Settings.UI.Lib.dll
