+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Libs/x64/Wox.Infrastructure.dll
