+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Libs/x64/Wox.Plugin.dll
