+++ Community.PowerToys.Run.Plugin.BrowserFavorite/Models/FavoriteType.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Davide Giacometti. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Community.PowerToys.Run.Plugin.BraveFavorite.Models
+namespace Community.PowerToys.Run.Plugin.BrowserFavorite.Models
 {
     public enum FavoriteType
     {
